AST Foundation
Implements as AST with three different node types: Idenfier, Literal, List. Identifiers can be used to reference functions and constants in a given runtime context. Literals can be used to represent literal values such as string, number, and boolean values. Lists can be used to represent a sequence of literal values or to represent an expression that results in a function call when evaluated.